Fixed precompilation issues
