+++ src/basetypes/BondGraph.jl
@@ -130,3 +130,14 @@ function getproperty(bgn::BondGraphNode, sym::Symbol)
 end
 
 exposed(bgn::BondGraphNode) = getnodes(bgn.bondgraph, SourceSensor)
+
+function port_info(t::Tuple{BondGraphNode,String})
+    pts = [n for n in nodes(t[1].bondgraph) if n isa SourceSensor]
+    for (i,c) in enumerate(pts)
+        if (c isa SourceSensor) && (t[2] == c.name)
+            return (t[1],i)
+        end
+    end
+    return error("Port $(t[2]) not found.")
+end
+port_info(t::Tuple{BondGraphNode,Symbol}) = port_info((t[1],string(t[2])))
